DL/Face Recognition/requirements.txt
keras==2.7.0
matplotlib==3.2.2
numpy==1.22.0
opencv_python==4.5.3.56
pandas==1.0.5
Pillow==9.1.1
tensorflow==2.9.1
